+++ pkg/metrics/http_server.go
@@ -1,16 +1,23 @@
 package metrics
 
 import (
-	"log"
 	"net/http"
 
+	"github.com/pkg/errors"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
+	"github.com/stackrox/rox/pkg/env"
+	"github.com/stackrox/rox/pkg/errorhelpers"
+	"github.com/stackrox/rox/pkg/logging"
 )
 
 const (
 	defaultAddress = ":9090"
 	metricsURLPath = "/metrics"
+)
+
+var (
+	log = logging.LoggerForModule()
 )
 
 // HTTPServer is a HTTP server for exporting Prometheus metrics.
@@ -20,16 +27,28 @@ type HTTPServer struct {
 	HandlerOpts promhttp.HandlerOpts
 }
 
-// NewDefaultHTTPServer creates and returns a new metrics http server with default settings.
+// NewDefaultHTTPServer creates and returns a new metrics http server with configured settings.
 func NewDefaultHTTPServer() *HTTPServer {
+	if err := env.ValidateMetricsSetting(); err != nil {
+		err = errorhelpers.PanicOnDevelopment(errors.Wrap(err, "invalid metrics port setting"))
+		log.Error(errors.Wrap(err, "metrics server is disabled"))
+	}
+	if !env.MetricsEnabled() {
+		log.Warn("Metrics server is disabled")
+		return nil
+	}
+
 	return &HTTPServer{
-		Address:  defaultAddress,
+		Address:  env.MetricsSetting.Setting(),
 		Gatherer: prometheus.DefaultGatherer,
 	}
 }
 
 // RunForever starts the HTTP server in the background.
 func (s *HTTPServer) RunForever() {
+	if s == nil {
+		return
+	}
 	mux := http.NewServeMux()
 	mux.Handle(metricsURLPath, promhttp.HandlerFor(s.Gatherer, s.HandlerOpts))
 	httpServer := &http.Server{